chore: Move fake safe mounter implementation to pkg/mounter (#13)
This change moves the fake mounter implementation from pkg/azuredisk to pkg/mounter. It formalizes it as FakeSafeMounter and uses the k8s.io/mount-utils fakes instead of the real implementation reducing and eliminating unnecessary dependencies for unit tests. (For example, it is no longer necessary to have csi-proxy when running the Windows unit tests.)

The clean up lays the foundation for the Node Provisioner library changes.