use mount-utils for resizing decision instead of context
Getting information whether or not a resize is required from request
context is not reliable.

The resize works for clones but a volume that was restored from a
snapshot fails to resize if the restore size is larger than the
origin/snapshot.

mount-utils now provides a resize checker that checks real volume size
and determines if the resize is needed reliably.